Removing directories, removing read-mes, minor renaming.
