--- qiskit/aqua/translators/data_providers/dataondemand/README.md
@@ -1,14 +1,0 @@
-# Qiskit Finance
-
-## Stock market data driver for NASDAQ Data on Demand
-
-NASDAQ is a major vendor of stock market data. It provides data not only for NASDAQ
-issues, but also for NYSE etc.
-
-This driver requires Data on Demand API Token.
-
-## Example query
-
-The data are obtained by running a query through the REST API.  
-```
-```
