--- qiskit/aqua/translators/data_providers/exchangedata/README.md
@@ -1,22 +1,0 @@
-# Qiskit Finance
-
-## Stock market data driver for Exchange Data International
-
-Exchange Data International is a major vendor of stock-market data. See 
-https://www.exchange-data.com/about_us.php#edi
-
-For samples of the data, please see:
-https://www.quandl.com/data/XSES-Singapore-Exchange-Prices
-https://www.quandl.com/data/XBER-Berlin-Stock-Exchange-Prices
-https://www.quandl.com/data/XPAR-Euronext-Paris-Stock-Prices/documentation
-
-This driver requires Quandl API Token.
-
-## Example query
-
-The data are obtained by running a query through quandl. See:
-https://docs.quandl.com/docs/parameters-2#section-times-series-parameters
-for details.
-
-```
-```
