--- qiskit/aqua/translators/data_providers/wikipedia/README.md
@@ -1,11 +1,0 @@
-# Qiskit Finance
-
-## Stock market data driver for Wikipedia
-
-Wikipedia contains stockmarket data, that are rather reliable up until 2018.
-
-## Example query
-
-The data are obtained by running a query through quandl.  
-```
-```
