+++ qiskit/aqua/translators/data_providers/wikipediadataprovider.py
